+++ core/src/main/kotlin/net/corda/core/ClockUtils.kt
@@ -1,9 +1,8 @@
-package net.corda.node.utilities
+package net.corda.core
 
 import co.paralleluniverse.fibers.Suspendable
 import co.paralleluniverse.strands.SettableFuture
 import com.google.common.util.concurrent.ListenableFuture
-import net.corda.core.then
 import rx.Observable
 import rx.Subscriber
 import rx.subscriptions.Subscriptions