+++ core/src/test/kotlin/net/corda/core/FiberBoxTest.kt
@@ -1,10 +1,8 @@
-package net.corda.node.utilities
+package net.corda.core
 
 import co.paralleluniverse.fibers.FiberExecutorScheduler
 import co.paralleluniverse.fibers.Suspendable
 import co.paralleluniverse.strands.Strand
-import net.corda.core.RetryableException
-import net.corda.core.getOrThrow
 import net.corda.testing.node.TestClock
 import org.junit.After
 import org.junit.Before