+++ node/src/main/kotlin/net/corda/node/services/persistence/NodeAttachmentService.kt
@@ -13,13 +13,13 @@ import net.corda.core.crypto.SecureHash
 import net.corda.core.div
 import net.corda.core.extractZipFile
 import net.corda.core.isDirectory
+import net.corda.core.node.services.AcceptsFileUpload
 import net.corda.core.node.services.AttachmentStorage
 import net.corda.core.serialization.CordaSerializable
 import net.corda.core.serialization.SerializationToken
 import net.corda.core.serialization.SerializeAsToken
 import net.corda.core.serialization.SerializeAsTokenContext
 import net.corda.core.utilities.loggerFor
-import net.corda.node.services.api.AcceptsFileUpload
 import net.corda.node.services.database.RequeryConfiguration
 import net.corda.node.services.persistence.schemas.AttachmentEntity
 import net.corda.node.services.persistence.schemas.Models