+++ samples/irs-demo/src/integration-test/kotlin/net/corda/irs/IRSDemoTest.kt
@@ -15,11 +15,11 @@ import net.corda.irs.contract.InterestRateSwap
 import net.corda.irs.utilities.postJson
 import net.corda.irs.utilities.putJson
 import net.corda.irs.utilities.uploadFile
-import net.corda.testing.driver.driver
 import net.corda.node.services.config.FullNodeConfiguration
 import net.corda.node.services.transactions.SimpleNotaryService
 import net.corda.nodeapi.User
 import net.corda.testing.IntegrationTestCategory
+import net.corda.testing.driver.driver
 import net.corda.testing.http.HttpApi
 import org.apache.commons.io.IOUtils
 import org.assertj.core.api.Assertions.assertThat