+++ samples/irs-demo/src/test/kotlin/net/corda/irs/plugin/IRSDemoPlugin.kt
@@ -1,0 +1,9 @@
+package net.corda.irs.plugin
+
+import net.corda.core.node.CordaPluginRegistry
+import net.corda.irs.api.IRSDemoAPI
+import java.util.function.Function
+
+class IRSDemoPlugin : CordaPluginRegistry() {
+    override val webApis = listOf(Function(::IRSDemoAPI))
+}